refactor: 계산 메소드와 입력 값 처리하는 메소드를 서로 다른 클래스로 분리
- StringCalculator.java: 실제 연산을 처리하는 로직을 담당하는 클래스

- ExpressionProcessor.java: 입력 값에 대한 유효성 검사 등을 담당하는 클래스 (추후 숫자, 기호 검사까지 추가할 예정)

- 위 변경 사항에 맞춰 테스트 코드도 분리
