+++ src/main/java/calculator/ExpressionProcessor.java
@@ -1,0 +1,13 @@
+package calculator;
+
+import java.util.Objects;
+
+public class ExpressionProcessor {
+
+    public static String[] splitExpression(String expression) {
+        if (Objects.isNull(expression) || expression.isEmpty()) {
+            throw new IllegalArgumentException("null 또는 빈 공백의 문자열입니다.");
+        }
+        return null;
+    }
+}
